requirements.txt
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0 